strlcat and strnstr

+++ ft_strlen.c
@@ -2,7 +2,7 @@
 
 size_t ft_strlen(const char *s)
 {
-  long long unsigned int i;
+  size_t i;
 
   i = 0;
   while(s[i])
